move general schema hash to common module (#5466)
* move general schema hash to common module
